pushing main

+++ BackEnd/src/main/java/com/revature/Project2/controller/UserController.java
@@ -70,14 +70,14 @@ public class UserController {
 
     @PostMapping("/deposit")
     @ResponseBody
-    public void deposit(@RequestParam Integer deposit, @RequestParam String username) {
+    public void deposit(@RequestParam double deposit, @RequestParam String username) {
         // System.out.println(deposit);
         // User user = new User()
         service.depositForUser(deposit, username);
     }
 
     @PostMapping("/withdrawal")
-    public ResponseEntity withdraw(@RequestParam Integer withdrawal, @RequestParam String username) {
+    public ResponseEntity withdraw(@RequestParam double withdrawal, @RequestParam String username) {
         if (service.userCanWithdraw(withdrawal, username)) {
             service.withdrawForUser(withdrawal, username);
             return new ResponseEntity<>("Successful Withdraw", HttpStatus.OK);
